barrier functions wrap unsafe behaviour
The safety documentation for the barrier functions just say "In your
own hands, this is hardware land!" because the programmer doesn't need
to check safety when adding a barrier.  Adding a memory barrier does
not break the borrow checker, cause undefined behaviour or alter
memory in any way.  Therefore, we can make dsb(), dmb() and isb() safe
wrappers around the asm code that write the intended assembly.